requirements.txt
firebase-admin
flask
python-dotenv
googlemaps
google-generativeai